[TOPI] Re-organise Arm(R) Cortex(R) CPU schedules to reflect CPU and Features
This patch splits the current schedules which have a mix of
architectural features into sub-modules so as to better ascertain which
features are in use. Specific schedules are moved to relevant folders and schedules which
work without features and select between features are kept at the root
level.
