+++ python/tvm/topi/arm_cpu/__init__.py
@@ -17,15 +17,14 @@
 # pylint: disable=wildcard-import
 """Schedule for ARM CPU"""
 
+from . import aprofile
+from . import mprofile
+
 from .conv1d import *
 from .conv2d import *
 from .depthwise_conv2d import *
 from .conv2d_transpose import *
 from .conv2d_int8 import *
-from . import conv2d_alter_op
-from .bitserial_conv2d import *
-from .bitserial_dense import *
-from .injective import *
+from .conv2d_alter_op import *
 from .group_conv2d import *
-from .pooling import *
-from .dense import *
+from .injective import *
