+++ python/tvm/topi/arm_cpu/aprofile/__init__.py
@@ -14,23 +14,7 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""Conv1D schedule for ARM CPU"""
-from __future__ import absolute_import as _abs
 
-from tvm import autotvm
-
-from .mprofile.dsp.conv1d import (
-    conv1d_nwc_dsp_compute,
-    conv1d_nwc_dsp_schedule,
-)
-
-
-@autotvm.register_topi_compute("conv1d_nwc_dsp.arm_cpu")
-def conv1d_nwc_dsp(cfg, data, kernel, strides, padding, dilation, out_dtype):
-    """Compute conv1d with v7e-m DSP instructions."""
-    return conv1d_nwc_dsp_compute(cfg, data, kernel, strides, padding, dilation, out_dtype)
-
-
-@autotvm.register_topi_schedule("conv1d_nwc_dsp.arm_cpu")
-def schedule_conv1d_nwc_dsp(cfg, outs):
-    return conv1d_nwc_dsp_schedule(cfg, outs)
+from . import asimd
+from . import dotprod
+from . import matmul
