+++ python/tvm/topi/arm_cpu/aprofile/asimd/bitserial_dense.py
@@ -14,17 +14,16 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-# pylint: disable=invalid-name, invalid-name, too-many-locals, too-many-arguments
 """Schedule for bitserial dense operator."""
 from __future__ import absolute_import as _abs
 import tvm
 from tvm import te
 from tvm import autotvm
 from tvm.topi.utils import get_const_tuple
-from .. import tag
-from .bitserial_conv2d import _intrin_popcount
-from ..nn.pad import pad
-from ..nn.bitserial_util import bitpack, binary_op_multiplier
+from .... import tag
+from .popcount import popcount
+from ....nn.pad import pad
+from ....nn.bitserial_util import bitpack, binary_op_multiplier
 
 
 @autotvm.register_topi_compute("bitserial_dense.arm_cpu")
@@ -179,7 +178,7 @@ def schedule_bitserial_dense(cfg, outs):
         nfactor = cfg["tile_y"].size[-1]
         kfactor = cfg["tile_k"].size[-1]
         if nfactor % 8 == 0:
-            pc = _intrin_popcount(nfactor, kfactor, WB, DB, unipolar)
+            pc = popcount(nfactor, kfactor, WB, DB, unipolar)
             s[output].tensorize(wb, pc)
 
         return s
