+++ python/tvm/topi/arm_cpu/aprofile/dotprod/__init__.py
@@ -14,20 +14,6 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-# pylint: disable=invalid-name, unused-variable, no-else-return, unused-argument, import-outside-toplevel
-"""Dense schedule for ARM CPU"""
 
-from tvm import autotvm
-from .mprofile.dsp.dense import dense_dsp_schedule, dense_dsp_compute
-
-
-@autotvm.register_topi_compute("dense_dsp.arm_cpu")
-def dense_dsp(cfg, data, weight, bias, out_dtype):
-    """Compute conv2d_nhwc with v7e-m DSP instructions."""
-    return dense_dsp_compute(cfg, data, weight, bias=bias, out_dtype=out_dtype)
-
-
-@autotvm.register_topi_schedule("dense_dsp.arm_cpu")
-def schedule_dense_dsp(cfg, outs):
-    """Create schedule for dense_dsp"""
-    return dense_dsp_schedule(cfg, outs)
+from .dotprod import *
+from .gemm import *
