+++ python/tvm/topi/arm_cpu/mprofile/__init__.py
@@ -15,3 +15,5 @@
 # specific language governing permissions and limitations
 # under the License.
 """Schedules specialized for cortex-m DSP instructions."""
+
+from . import dsp
