+++ python/tvm/topi/arm_cpu/mprofile/dsp/__init__.py
@@ -14,3 +14,8 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+
+from .dense import *
+from .conv1d import *
+from .conv2d import *
+from .pool import *
